Improved save file compatibility handling.
* Add an internal version counter, starting at 1.
  Said value should be increased any time compatibility is broken, i.e.
  attempting to load an old save segfaults.
* Check this value when loading a save file.
* If incompatible, prompt whether the file should be deleted.
  If yes, start a new game.

This breaks saves but seeing how the map memory change (storing item
colours) already did, I don't think this is much of a problem.

The prompt is a *huge* improvement considering the last time 
compatibility broke and I had an unfinished game, I had no idea where 
the file was saved (didn't help that the folder was hidden, either)
and had to search the source to find its location and delete it
before I could continue playing. :( 


git-svn-id: https://nlarn.svn.sourceforge.net/svnroot/nlarn/trunk@504 561b6253-61b6-49b5-99ee-40376865e0bd

+++ src/nlarn.c
@@ -52,14 +52,41 @@ int main(int argc, char *argv[])
     g_free(userdir);
 
     /* find save file */
+    gboolean loaded = FALSE;
+    gboolean display_initialised = FALSE;
     if (g_file_test(save_file_name, G_FILE_TEST_EXISTS | G_FILE_TEST_IS_REGULAR))
     {
         /* restore savegame */
-        game_load(save_file_name, argc, argv);
+        loaded = game_load(save_file_name, argc, argv);
 
-        /* delete save file */
-        g_unlink(save_file_name);
+        if (!loaded)
+        {
+            /* initialize display */
+            display_init();
 
+            /* call display_shutdown when terminating the game */
+            atexit(display_shutdown);
+
+            display_draw();
+            display_initialised = TRUE;
+
+            if (display_get_yesno("Saved game could not be loaded. " \
+                                  "Delete and start new game?",
+                                  NULL, NULL))
+            {
+                /* delete save file */
+                g_unlink(save_file_name);
+            }
+            else
+            {
+                g_printerr("Save file not compatible to current version.\n");
+                exit(EXIT_FAILURE);
+            }
+        }
+    }
+
+    if (loaded)
+    {
         /* refresh FOV */
         player_update_fov(nlarn->p);
     }
@@ -75,13 +102,16 @@ int main(int argc, char *argv[])
         scroll_mapping(nlarn->p, NULL);
     }
 
-    /* initialize display */
-    display_init();
+    if (!display_initialised)
+    {
+        /* initialize display */
+        display_init();
 
-    /* call display_shutdown when terminating the game */
-    atexit(display_shutdown);
+        /* call display_shutdown when terminating the game */
+        atexit(display_shutdown);
 
-    display_draw();
+        display_draw();
+    }
 
     /* check if mesgfile exists */
     if (!g_file_get_contents(game_mesgfile(nlarn), &strbuf, NULL, NULL))
